finished article factories

+++ InventoryManagement/Models/Articles/WomenShoe.cs
@@ -8,7 +8,7 @@ public class WomenShoe : Shoe
     /// <summary>
     /// Gets or sets the height of the heel in cm
     /// </summary>
-    public double HeelHeight { get; set; }
+    public decimal HeelHeight { get; set; }
 
     /// <summary>
     /// Creates a new inventory item that represents a shoe for women
@@ -17,6 +17,6 @@ public class WomenShoe : Shoe
     public WomenShoe(int id) 
         : base(id)
     {
-        HeelHeight = 0d;
+        HeelHeight = 0m;
     }
 }
